overlooked that it also looked up 'duration', casted that also to long to try and fix #2134

+++ core/src/main/java/de/danoeh/antennapod/core/service/playback/PlaybackService.java
@@ -1331,7 +1331,7 @@ public class PlaybackService extends MediaBrowserServiceCompat {
             if (queue != null) {
                 i.putExtra("ListSize", queue.size());
             }
-            i.putExtra("duration", info.playable.getDuration());
+            i.putExtra("duration", (long)info.playable.getDuration());
             i.putExtra("position", (long)info.playable.getPosition());
             sendBroadcast(i);
         }
